Implémentation du circuit d'authentification

+++ gameref-back/src/main/java/fr/orsys/groupe3/gamerefback/controller/GameRestController.java
@@ -1,8 +1,10 @@
 package fr.orsys.groupe3.gamerefback.controller;
 
 import fr.orsys.groupe3.gamerefback.business.Game;
+import fr.orsys.groupe3.gamerefback.business.Moderator;
 import fr.orsys.groupe3.gamerefback.dto.GameDto;
 import fr.orsys.groupe3.gamerefback.exception.NotFoundException;
+import fr.orsys.groupe3.gamerefback.exception.SecurityException;
 import fr.orsys.groupe3.gamerefback.service.GameService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.data.domain.Pageable;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
 import java.util.*;
@@ -20,21 +23,37 @@ import java.util.*;
 
 @RestController
 @RequestMapping("/api")
-@CrossOrigin(origins = "*", maxAge = 3600)
 @AllArgsConstructor
+@CrossOrigin(origins = "http://localhost:4200", maxAge = 3600, allowCredentials = "true")
 public class GameRestController {
 
     private GameService gameService;
 
+    private HttpSession httpSession;
+
     // Ajout d'un jeu
     @PostMapping("/game")
-    public Game addGame(@RequestBody GameDto dto) throws NotFoundException {
-        return gameService.createGame(dto, null);
+    public Game addGame(@RequestBody GameDto dto) throws NotFoundException, SecurityException {
+        Object user = httpSession.getAttribute("user");
+        if (user == null) {
+            throw new SecurityException("Session expired, please login as a moderator and try again");
+        }
+        if (!(user instanceof Moderator)) {
+            throw new SecurityException("Forbidden");
+        }
+        return gameService.createGame(dto, (Moderator) user);
     }
 
     // Suppression d'un jeu
     @DeleteMapping("game/{id}")
-    public Game deleteGame(@PathVariable("id") long id) throws NotFoundException {
+    public Game deleteGame(@PathVariable("id") long id) throws NotFoundException, SecurityException {
+        Object user = httpSession.getAttribute("user");
+        if (user == null) {
+            throw new SecurityException("Session expired, please login as a moderator and try again");
+        }
+        if (!(user instanceof Moderator)) {
+            throw new SecurityException("Forbidden");
+        }
         return gameService.deleteGame(id);
     }
 
@@ -44,7 +63,14 @@ public class GameRestController {
     }
 
     @PatchMapping("/game/{id}")
-    public Game updateGame(@PathVariable Long id, @RequestBody GameDto dto) throws NotFoundException {
+    public Game updateGame(@PathVariable Long id, @RequestBody GameDto dto) throws NotFoundException, SecurityException {
+        Object user = httpSession.getAttribute("user");
+        if (user == null) {
+            throw new SecurityException("Session expired, please login as a moderator and try again");
+        }
+        if (!(user instanceof Moderator)) {
+            throw new SecurityException("Forbidden");
+        }
         return gameService.updateGame(id, dto);
 
     }
@@ -74,5 +100,11 @@ public class GameRestController {
             }
         }
         return errors;
+    }
+
+    @ExceptionHandler(SecurityException.class)
+    @ResponseStatus(code=HttpStatus.FORBIDDEN)
+    public Map<String, List<String>> handleSecurityException(SecurityException exception) {
+        return Map.of("access", List.of(exception.getMessage()));
     }
 }
